Examples(Lang Tour): adjust query selector, slight code refactor, README update
- Contributed to first two items of #437
- Also moved `lib/language_tour/util/*` to `lib/util/`.
