+++ examples/lib/language_tour/classes/enum.dart
@@ -1,4 +1,4 @@
-import '../util/print.dart';
+import '../../util/print.dart';
 
 // #docregion enum
 enum Color { red, green, blue }
