+++ examples/lib/language_tour/classes/logger_with_main.dart
@@ -1,4 +1,4 @@
-import '../util/print.dart';
+import '../../util/print.dart';
 
 // #docregion
 class Logger {
