+++ examples/lib/language_tour/classes/point_with_distance_field.dart
@@ -1,4 +1,4 @@
-import '../util/print.dart';
+import '../../util/print.dart';
 // #docregion
 import 'dart:math';
