+++ examples/lib/language_tour/classes/proxy.dart
@@ -2,7 +2,7 @@
 // Dart 2.0: hint • 'proxy' is deprecated and shouldn't be used
 // ignore_for_file: deprecated_member_use
 
-import '../util/print.dart';
+import '../../util/print.dart';
 
 void main() {
   dynamic a = new A();
