+++ examples/lib/language_tour/classes/proxy_alt.dart
@@ -1,4 +1,4 @@
-import '../util/print.dart';
+import '../../util/print.dart';
 
 void main() {
   dynamic a = new A();
