+++ examples/lib/language_tour/generics/base_class.dart
@@ -1,4 +1,4 @@
-import '../util/print.dart';
+import '../../util/print.dart';
 
 class SomeBaseClass {}
