+++ examples/lib/util/logging_printer.dart
