+++ examples/lib/util/print.dart
