+++ examples/test/language_tour/basic_test.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: type_annotate_public_apis
 import 'package:test/test.dart';
-import 'package:examples/language_tour/util/logging_printer.dart';
-import 'package:examples/language_tour/util/print.dart';
+import 'package:examples/util/logging_printer.dart';
+import 'package:examples/util/print.dart';
 
 void main() {
   final printLog = PrintLog.it;
