+++ examples/test/language_tour/control_flow_test.dart
@@ -1,6 +1,6 @@
 import 'package:test/test.dart';
-import 'package:examples/language_tour/util/logging_printer.dart';
-import 'package:examples/language_tour/util/print.dart';
+import 'package:examples/util/logging_printer.dart';
+import 'package:examples/util/print.dart';
 
 void main() {
   final printLog = PrintLog.it;
