+++ examples/test/language_tour/functions_test.dart
@@ -3,8 +3,8 @@ import 'package:test/test.dart';
 
 import 'package:examples/language_tour/function_equality.dart'
     as function_equality;
-import 'package:examples/language_tour/util/logging_printer.dart';
-import 'package:examples/language_tour/util/print.dart';
+import 'package:examples/util/logging_printer.dart';
+import 'package:examples/util/print.dart';
 
 void main() {
   final printLog = PrintLog.it;
