+++ examples/test/language_tour/generics_test.dart
@@ -2,8 +2,8 @@
 // #docplaster
 import 'package:test/test.dart';
 import 'package:examples/language_tour/generics/base_class.dart' as base_class;
-import 'package:examples/language_tour/util/logging_printer.dart';
-import 'package:examples/language_tour/util/print.dart';
+import 'package:examples/util/logging_printer.dart';
+import 'package:examples/util/print.dart';
 
 final Matcher throwsATypeError = throwsA(new isInstanceOf<TypeError>());
